Speed up Timeline render function by a few times
Among other things, replaces some uses of Object with Map as getting a
value from an Object is extremely slow when the key is a Date or Moment
object. Also, according to the MDN docs at
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map#Objects_and_maps_compared,
a Map may perform better in scenarios involving frequent addition and
removal of key pairs.