Remove results page slider
The slider was implemented during the FB Hackathon, at commit
https://github.com/WotFace/ShowFace/commit/c34a63988baec0d111d9f4c645e2ce17b777e766.
However, we don't really need it (it was a workaround for the lack of a
respondent filter), and the code to support this complicates the
ShowRespond and Timeline components too much, impeding progress on the
Timeline component. We can add this back if necessary when the app is
further developed.